--- specs/wdio.spec1.js
@@ -1,129 +1,0 @@
-const { expect } = require('chai');
-const { App } = require('../src/pages');
-const rundomNumber = () => Date.now();
-
-const app = new App();
-
-describe('Registration:', function () {
-  beforeEach(async function () {
-    await browser.setWindowSize(1440, 960);
-    await browser.url('/sign-up');
-  });
-
-  afterEach(async function () {
-    await browser.reloadSession();
-  });
-
-  xit('should be able to register doctor', async function () {
-    await app.authPage.register({
-      name: `John${rundomNumber()}`,
-      surname: 'Doctor',
-      email: `marcus${rundomNumber()}@gmail.com`,
-      password: 'Pa55word',
-      phone: '380999999',
-      birthDate: '11/11/2000',
-      status: 'doctor',
-      gender: 'female',
-    });
-
-    await browser.waitUntil(
-      async function () {
-        const url = await browser.getUrl();
-        return url === 'http://46.101.234.121/doctors';
-      },
-      { timeout: 5000 },
-    );
-
-    const url = await browser.getUrl();
-    expect(url).to.be.eql('http://46.101.234.121/doctors');
-  });
-
-  xit('should be able to register patient', async function () {
-    await app.authPage.register({
-      name: `John${rundomNumber()}`,
-      surname: 'Patient',
-      email: `marcus${rundomNumber()}@gmail.com`,
-      password: 'Pa55word',
-      phone: '380999999',
-      birthDate: '11/11/2000',
-      status: 'patient',
-      gender: 'male',
-    });
-
-    await browser.waitUntil(
-      async function () {
-        const url = await browser.getUrl();
-        return url === 'http://46.101.234.121/doctors';
-      },
-      { timeout: 5000 },
-    );
-
-    const url = await browser.getUrl();
-    expect(url).to.be.eql('http://46.101.234.121/doctors');
-  });
- 
-  const { expect } = require('chai');
-const { App } = require('../src/pages');
-const rundomNumber = () => Date.now();
-
-const app = new App();
-
-describe('Sign in:', function () {
-  beforeEach(async function () {
-    await browser.setWindowSize(1440, 960);
-    await browser.url('/sign-in');
-  });
-
-  afterEach(async function () {
-    await browser.reloadSession();
-  });
-
-  it('should be able to log in', async function () {
-    await app.authPage.login({
-     email: 'john_admin2@admin.com',
-     password: 'Pa55word',
-    });
-
-    await browser.waitUntil(
-      async function () {
-        const url = await browser.getUrl();
-        return url === 'http://46.101.234.121/doctors';
-      },
-      { timeout: 5000 },
-    );
-
-    const url = await browser.getUrl();
-    expect(url).to.be.eql('http://46.101.234.121/doctors');
-  });
-
-  describe('Autorization with entering invalid data:', function () {
-    beforeEach(async function () {
-      await browser.setWindowSize(1440, 960);
-      await browser.url('/sign-in');
-    });
-  
-    afterEach(async function () {
-      await browser.reloadSession();
-    });
-
-  xit('should not be able to log in', async function () {
-    await app.authPage.login({
-     email: 'john_admin2@admin.com',
-     password: 'Pa44word',
-    });
-
-    await browser.waitUntil(
-      async function () {
-        const url = await browser.getUrl();
-        return url === '/sign-in';
-      },
-      { timeout: 5000 },
-    );
-
-    const errorMessage = $('div.rrt-text=Password is incorrect');
-    expect(errorMessage).toBeDisplayed();
-
-  });
-  });
-});
-});
